chaewsstore-admin/src/main/java/com/chaewsstore/admin/config/security/handler/JwtAccessDeniedHandler.java
package com.chaewsstore.admin.config.security.handler;

import static com.chaewsstore.core.infra.exception.JwtErrorCode.ACCESS_DENIED;
import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.globalutils.response.ErrorResponse;
import com.globalutils.exception.StatusCode;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.IOException;
import lombok.RequiredArgsConstructor;
import org.springframework.security.access.AccessDeniedException;
import org.springframework.security.web.access.AccessDeniedHandler;

@RequiredArgsConstructor
public class JwtAccessDeniedHandler implements AccessDeniedHandler {

    private final ObjectMapper objectMapper;

    @Override
    public void handle(HttpServletRequest request, HttpServletResponse response,
        AccessDeniedException accessDeniedException) throws IOException {

        response.setStatus(StatusCode.FORBIDDEN.getCode());
        response.setContentType(APPLICATION_JSON_VALUE);
        response.setCharacterEncoding("UTF-8");
        response.getWriter().write(objectMapper.writeValueAsString(
            ErrorResponse.of(ACCESS_DENIED, accessDeniedException.getMessage())));
    }
}
