chaewsstore-app/src/test/java/com/chaewsstore/app/common/security/handler/JwtAccessDeniedHandlerTest.java
package com.chaewsstore.app.common.security.handler;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

import com.chaewsstore.app.config.security.handler.JwtAccessDeniedHandler;
import com.fasterxml.jackson.databind.ObjectMapper;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.IOException;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.security.access.AccessDeniedException;

@ExtendWith(MockitoExtension.class)
class JwtAccessDeniedHandlerTest {

    @Mock
    private ObjectMapper objectMapper;

    @InjectMocks
    private JwtAccessDeniedHandler jwtAccessDeniedHandler;

    @Test
    void testHandle() throws IOException {
        HttpServletRequest request = new MockHttpServletRequest();
        HttpServletResponse response = new MockHttpServletResponse();
        AccessDeniedException accessDeniedException = new AccessDeniedException("Access denied");

        String errorResponseJson = "{\"code\":\"ACCESS_DENIED\",\"message\":\"Access denied\"}";
        when(objectMapper.writeValueAsString(any())).thenReturn(errorResponseJson);

        jwtAccessDeniedHandler.handle(request, response, accessDeniedException);

        assertEquals(HttpServletResponse.SC_FORBIDDEN, response.getStatus());
        assertEquals("application/json;charset=UTF-8", response.getContentType());
        assertEquals("UTF-8", response.getCharacterEncoding());
    }
}
